Endpoints in config, api moved to dedicated folder
