+++ src/App.tsx
@@ -29,7 +29,7 @@ import '@ionic/react/css/display.css';
 /* Theme variables */
 import './theme/variables.css';
 import LockdropStatPage from './pages/LockdropStatPage';
-import Api from './contexts/Api';
+import Api from './api/Api';
 import Web3Api from './contexts/Web3Api';
 
 const App: React.FC = () => {
