+++ src/components/RealtimeLockdrop/ClaimStatus.tsx
@@ -32,7 +32,7 @@ import CopyMessageBox from '../CopyMessageBox';
 import ClaimItem from './ClaimableItem';
 import BigNumber from 'bignumber.js';
 import moment from 'moment';
-import { useApi } from '../../contexts/Api';
+import { useApi } from '../../api/Api';
 import useChainInfo from '../../helpers/useChainInfo';
 
 interface Props {
