+++ src/config/endpoints.ts
@@ -1,0 +1,32 @@
+import { PlasmNetwork } from '../helpers/plasmUtils';
+
+const DEFAULT_ENDPOINT_INDEX = 2;
+
+interface Endpoint {
+    network: PlasmNetwork;
+    address: string;
+}
+
+const endpoints: Endpoint[] = [
+    {
+        network: PlasmNetwork.Local,
+        address: 'ws://127.0.0.1:9944',
+    },
+    {
+        network: PlasmNetwork.Dusty,
+        address: 'wss://rpc.dusty.plasmnet.io/',
+    },
+    {
+        network: PlasmNetwork.Main,
+        address: 'wss://rpc.plasmnet.io',
+    },
+];
+
+/**
+ * gets endpoint url for a given network
+ * @param network the network
+ */
+export function getNetworkEndpoint(network?: PlasmNetwork) {
+    const endpoint = endpoints.find(x => x.network === network) || endpoints[DEFAULT_ENDPOINT_INDEX];
+    return endpoint.address;
+}
